Incorporate @afeld's feedback from #5
Namely, split single-line function definitions across lines and extracted a
new RequestHelper class to trim down the test/index.js file.
